Fixing inconsistent response of pipeline list endpoint
During my work on https://github.com/bob-cd/wendy/issues/10
I determined a supposable bug in the API of Bob. It does not
return a response consistent with the other list endpoints.
The fix is to use the same response function from the util
namespace.

I refactored the get-pipelines function because I had the problem
that it was not clear which result was used in the answer. And
with the nested let it was not easy to read. Furthermore there
was a return additionally to the response and that I did not
understand.

- Fixes #73
- Relates to https://github.com/bob-cd/wendy/issues/10
- Unfortunately it breaks the API to get into a consistent state