Improved Bobs configuration (#34)
Bobs configuration now is validated by schema and has its own
namespace. All configuration is stored in a config variable
that is accessible and only written to at startup. People can
now use a .bob.conf in their home directory to configure Bob
with an edn-file.

The file is merged last so all configuration written in the
configuration file overwrites the other configurations. The
documentation is adapted to show how to write this file. When
a file is provided but cannot be parsed or merged it does
not affect the use of the other configuration parameters.

- Closes #34
- adapted docker-compose files
- Bob's server port is configurable
- compose-files updated with new vars