feat: event streaming
This commit adds streaming of events from runners to
the apiservers via RabbitMQ streams and streaming of
events from apiservers to connected clients via SSE.

This was necessary to enable building dynamic
frontends that can show updates on pipelines without
reloading. Please see the ADR for background info
and the issue #119 and PR #126.

- Closes #119